WIP move REST API types into workspaces
