+++ packages/types-rest-api-ghes-3.1/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@octokit-next/types-rest-api-ghes-3.1",
+  "publishConfig": {
+    "access": "public"
+  },
+  "type": "module",
+  "version": "0.0.0-development",
+  "description": "Types for GHES 3.1 REST API requests and responses",
+  "types": "./index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/octokit/octokit-next.js.git",
+    "directory": "packages/types-rest-api-ghes-3.1"
+  },
+  "author": "Gregor Martynus (https://twitter.com/gr2m)",
+  "license": "MIT"
+}
