+++ packages/types-rest-api-github.com/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@octokit-next/types-rest-api-github.com",
+  "publishConfig": {
+    "access": "public"
+  },
+  "type": "module",
+  "version": "0.0.0-development",
+  "description": "Types for github.com REST API requests and responses",
+  "types": "./index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/octokit/octokit-next.js.git",
+    "directory": "packages/types-rest-api-github.com"
+  },
+  "author": "Gregor Martynus (https://twitter.com/gr2m)",
+  "license": "MIT"
+}
