+++ tests/ghes-3.0/index.d.ts
@@ -1,6 +1,6 @@
 import { ExtendOctokitWith, Octokit } from "../../index.js";
 
-import "../../api-versions-types/ghes-3.0";
+import "@octokit-next/types-rest-api-ghes-3.0";
 
 export const OctokitGhes30: ExtendOctokitWith<
   Octokit<"ghes-3.0">,
